+++ packages/components/organisms/f-status-banner/src/constants.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export const BASE_HERO_URL = 'https://d30v2pzvrfyzpo.cloudfront.net/status-banner/';