+++ packages/components/organisms/f-status-banner/CHANGELOG.md
@@ -3,6 +3,17 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.0.0
+------------------------------
+*February 12, 2021*
+
+### Added
+- `PageBanner.vue`, `MainBannerContainer.vue` components.
+- Copy for each tenant.
+- Test to cover changes.
+- `constants.js` file.
+- `f-searchbox`, `f-header` & `f-footer` components.
+
 
 v0.1.0
 ------------------------------